[package-manager] Split out npm package manager
